+++ src/__init__.py
@@ -3,6 +3,7 @@
 from flask import Flask
 from werkzeug.middleware.proxy_fix import ProxyFix
 
+import src.index.routes
 import src.topics.routes
 import src.users.routes
 from src import routes
@@ -12,6 +13,7 @@ app = Flask(__name__)
 app.register_blueprint(routes.bp)
 app.register_blueprint(src.users.routes.bp)
 app.register_blueprint(src.topics.routes.bp)
+app.register_blueprint(src.index.routes.bp)
 app.config.from_object(Config)
