+++ src/index/__init__.py
@@ -1,0 +1,1 @@
+"""'index' package's initialization module."""
